daq_start function correction, and added daq command error detection for 2nd input being wrong
daq_start has some "or-ing" of the statuses that causes it not to behave properly from time to time. So I fixed that to make sure that it only detects the proper errors to make daq status reflect daq true state by removing the "or-ring" of statuses and replacing them with equal signs.