2025 10 01 LW Fixed ff_memalloc and ff_memfree (see desc)
This fix is technically a kludge but it has worked with my testing so far. We have very little remaining RAM to work with so I create a single static buffer of the required size for LFN operations. Because we are now using FF_FS_REENTRANT, calls to ff_memalloc and ff_memfree will be protected by a mutex, so we can get away with reusing the same single static buffer. Ideally we would use dynamic memory allocation and freeing for this, with the capability to allocate multiple buffers at once, but we simply don't have enough RAM left to accommodate more than one minimum size buffer.